# config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Ecommerce",
    "version": "1.0.0",
    "description": "Documentación de la API para la tienda Ecommerce"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Servidor de Desarrollo"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID único del pedido",
            "example": 1
          },
          "usuarioId": {
            "type": "integer",
            "description": "ID del usuario que realizó el pedido",
            "example": 1
          },
          "total": {
            "type": "number",
            "format": "float",
            "description": "Total del pedido",
            "example": 599.98
          },
          "direccionEnvio": {
            "type": "string",
            "description": "Dirección de envío",
            "example": "Calle Falsa 123, Ciudad, País"
          },
          "metodoPago": {
            "type": "string",
            "description": "Método de pago utilizado",
            "example": "tarjeta"
          },
          "estado": {
            "type": "string",
            "description": "Estado del pedido",
            "example": "pendiente"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación del pedido",
            "example": "2023-10-01T12:34:56.789Z"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/orders/checkout": {
      "post": {
        "summary": "Realizar checkout y crear un pedido",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "direccionEnvio",
                  "metodoPago"
                ],
                "properties": {
                  "direccionEnvio": {
                    "type": "string",
                    "description": "Dirección de envío"
                  },
                  "metodoPago": {
                    "type": "string",
                    "description": "Método de pago (ej. tarjeta, paypal)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "order": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Carrito vacío o datos inválidos"
          },
          "500": {
            "description": "Error al procesar el pedido"
          }
        }
      }
    },
    "/orders": {
      "get": {
        "summary": "Obtener historial de pedidos",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado"
          },
          "500": {
            "description": "Error al obtener los pedidos"
          }
        }
      }
    },
...
  },
...
}